Default dev server binding to localhost
Binding the dev server to all net interfaces on 0.0.0.0 by default presents a danger to workstations that are improperly firewalled, potentially exposing the server publicly when it isn't intended. For development purposes, local 127.0.0.1 only should be the default as that will match developer expectations. In circumstances where the port does need to be opened for access by an external viewer, devs can set `GEOPORTAL_SERVER_BIND_INTERFACE`.